tried kernel is integer now bot wrong tracking result

+++ ESLab_Assign02_Group11/code/meanshift.h
@@ -26,7 +26,7 @@ public:
     MeanShift();
     void Init_target_frame(const cv::Mat &frame,const cv::Rect &rect);
     float & kernel_elem(int row, int col, int height, int width);
-    float Epanechnikov_kernel(cv::Mat &kernel, int h, int w);
+    int Epanechnikov_kernel(cv::Mat &kernel, int h, int w);
     cv::Mat pdf_representation(const cv::Mat &frame,const cv::Rect &rect, const float init);
     cv::Mat CalWeight(const cv::Mat &frame, cv::Mat &target_model, cv::Mat &target_candidate, cv::Rect &rec);
     cv::Mat CalWeight_opt(const cv::Mat &frame, cv::Mat &target_model, cv::Mat &target_candidate, cv::Rect &rec);
